ssh: config directive to modify the local environment
This provides a mechanism to attach arbitrary configure options into the
ssh_config file and use them from the LocalCommand and ProxyCommand.

Examples:

    # set FOO to foo
    LocalEnvMod FOO = foo

    # append bar to FOO with default separator ","
    LocalEnvMod FOO += bar

    # unset FOO
    LocalEnvMod FOO =

    # append foo to BAR with separator ":", if BAR is empty no separator will
    # be used
    LocalEnvMod BAR +:= foo

    # prepend baz to BAR with separator ":"
    LocalEnvMod BAR %:= baz

Currently any chararacter can be used as separator.

My intended use case for this is to automount arbitrary remote directories
via sshfs based on the host by the mux master via LocalCommand. I have a
default of 'mount the remote home at ~/Remotes/<host>' but I can add more
mount directives based on individual hosts.

The LocalEnvMod directive is cumulative and are exectuted in one file in order
(ie. from top to buttom), but in reverse order of file parsing. That is
directives in ~/.ssh/config are executed after /etc/ssh/ssh_config but
command line directives (-oLocalEnvMod=) will be executed last.

If there are security concerns what variables are allowed to change, we could
use a blacklist (HOME, SHELL, ... come to mind).

---

v2: This time with my own fix for the %n expension.

# Conflicts:
#	readconf.c